fix tests
- use ts-jest
- move tests to tests-folder
- remove react-test-renderer for being redundant
